Not sure if this works, but it's different.
Couldnt test properly because woo I love having to do study hall in a lab, but it may work because it is differnt then what we had before. I think we need to set up a timer that will load the plot and then the song, but getting to run the plot and the song at the same time would require this...if not possible have the timer sleep long enough to build the plot and then pull up the song after in one sleep rotation. Still cloudy...